Fix: When job should be preserved update job shouldn't return True
When the script found an obsolete build job (either 32-bit or MinGW) the update script returned 'True' as a signal for the remote update script, that there are changes which need to be pushed to remote

Even when the

# bincrafters-conventions:preserve-build-job

tag was used to NOT delete this build job.

The result is that the job update scripts signals 'I made a change' without applying any change which could be pushed. The result can be an empty commit

Fix: Only return 'True' when a legacy build job was found and the preserve build tag was NOT found.