Cron: delete branches which got merged via a squashed commit
This only works if the commit message contains a reference to the PR in its message. Otherwise the commit contains zero information about the PR it is coming from.

A squashed commit is essential a completely new commit with now direct relation to its source branch.

Hence the reason why our current approach for deleting merged branches stopped working.